RustFFT Documentation

Compute FFTs in Julia using RustFFT. Some parts of this documentation have been quoted from the RustFFT docs.

RustFFT is a high-performance, SIMD-accelerated FFT library written in pure Rust. It can compute FFTs of any size, including prime-number sizes, in O(nlogn) time.

Usage

Forward FFT:

using RustFFT

planner64 = RustFFT.FftPlanner64()
instance = RustFFT.plan_fft_forward(planner64, UInt(1))
data = complex([1.0])
RustFFT.fft!(instance, data)
@assert data[1] ≈ 1.0

Inverse FFT:

using RustFFT

planner64 = RustFFT.FftPlanner64()
instance = RustFFT.plan_fft_inverse(planner64, UInt(1))
data = complex([1.0])
RustFFT.fft!(instance, data)
@assert data[1] ≈ 1.0

Note that RustFFT does not normalize outputs:

Callers must manually normalize the results by scaling each element by 1/len().sqrt(). Multiple normalization steps can be merged into one via pairwise multiplication, so when doing a forward FFT followed by an inverse callers can normalize once by scaling each element by 1/len()

It's currently not possible to choose the specific algorithm that will be used to compute the transform.

RustFFT.FftInstance32 — Type

FftInstance32

A planned FFT instance that can compute either forward or inverse FFTs of Vector{Complex{Float32}} data whose length is an integer multiple of the planned length.

RustFFT.FftInstance64 — Type

FftInstance64

A planned FFT instance that can compute either forward or inverse FFTs of Vector{Complex{Float64}} data whose length is an integer multiple of the planned length.

RustFFT.FftPlanner32 — Type

FftPlanner32

A planner for forward and inverse FFTs of Vector{Complex{Float32}} data. A new planner can be created by calling the zero-argument constructor.

RustFFT.FftPlanner64 — Type

FftPlanner64

A planner for forward and inverse FFTs of Vector{Complex{Float64}} data. A new planner can be created by calling the zero-argument constructor.

RustFFT.fft! — Function

fft!(instance, data)

Computes the forward or inverse FFT of the data in-place. instance must be either a FftInstance32 or a FftInstance64. data must be either a Vector{Complex{Float32}} or a Vector{Complex{Float64}}, the width must match the that of the provided instance, its length must be an integer multiple the length of the instance.

RustFFT.fft_async! — Function

fft_async!(instance, data)

Computes the forward or inverse FFT of the data in-place. The transform is computed by a background thread, this function waits for that computation to be finished before returning. See RustFFT.fft! for more info.

RustFFT.plan_fft — Function

plan_fft(planner, direction::Symbol, len::UInt)

Plan either a forward or an inverse FFT of length len. Returns either an FftInstance32 or FftInstance64 depending on the provided planner, which must be either an FftPlanner32 or FftPlanner64. The direction must be either :forward or :inverse

RustFFT.plan_fft_forward — Function

plan_fft_forward(planner, len::UInt)

Plan a forward FFT of length len. Returns either an FftInstance32 or FftInstance64 depending on the provided planner, which must be either an FftPlanner32 or FftPlanner64.

RustFFT.plan_fft_inverse — Function

plan_fft_inverse(planner, len::UInt)

Plan an inverse FFT of length len. Returns either an FftInstance32 or FftInstance64 depending on the provided planner, which must be either an FftPlanner32 or FftPlanner64.